Feature: model_name overridden in form_object module
The model_name class method is part of the ActiveModel::Naming module.

Typically Form Objects want to override the self.model_name method,
as this method is used by form_for to determine the URL that the form is
submitted to.

The FormObject module uses the resource part of the concrete form-object
Class name as the name attribute of the ActiveModel::Name instance
returned from model_name.