Feature: Add strong_parameters class method
In rails 4 the strong_parameters gem is used as a method of protecting
the attributes that can be passed through in the creation of a model.
The usual way of doing this is defining a method on your controller
however this is not the preferred location for defining the
strong_parameters when using form_objects. To prevent having to define a
boilerplate `x_params` method on every form object a `strong_parameters`
class method has been added.