fix(travis): configure sqlite to not be in-memory
Travis was configured to use an sqlite in-memory database for the test
environment. This was causing database dependent tests to fail as the in
memory database did not appear to be reliably migrated before tests.

This has been updated to configure sqlite to use a disc file, and the
database dependent tests now reliably pass.