fix(generator): create/destroy shared examples when required
Previously, the generator was greedy
(https://github.com/mattfreer/active_model_form_objects/issues/3),
however, now the generator only creates the shared example specs if
there are none and destroys them if it is destroying the last active
model form object.

The method hooks into the behavior on the class and performs a task when
it is set to `invoke` or `revoke`
(http://factore.ca/blog/126-reversible-rails-generators).